requirements.txt
streamlit==1.54.0
streamlit-autorefresh==1.0.1
pydantic==2.12.5
sqlalchemy==2.0.46
pyyaml==6.0
python-dotenv==1.0.0
yfinance==1.1.0
feedparser==6.0.12
anthropic==0.79.0
plotly==6.5.2
beautifulsoup4==4.12.2
requests==2.32.5
aiohttp==3.8.4
kiteconnect==5.0.1
nsepython==2.97
numpy==2.2.6
pandas==2.3.3
# praw==7.7.0              # Reddit sentiment (optional)
# youtube-transcript-api==0.6.0  # YouTube sentiment (optional)
